% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fetchCurrentWeather}
\alias{fetchCurrentWeather}
\title{fetchCurrentWeather}
\usage{
fetchCurrentWeather(queryCity)
}
\arguments{
\item{queryCity}{A String}
}
\value{
A DataFrame
}
\description{
fetchCurrentWeather
}
\examples{
fetchCurrentWeather("Pakistan")

}
