% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{fetchWeatherData}
\alias{fetchWeatherData}
\title{fetchWeatherData}
\usage{
fetchWeatherData(queryCity)
}
\arguments{
\item{queryCity}{A String}
}
\value{
A DataFrame
}
\description{
fetchWeatherData
}
\examples{
fetchWeatherData("pakistan")

}
